integration of new mutation density to deepCSA (#339)
* connect mut_density with pipeline's utils

* apply changes from PR #338
