+++ bin/utils_impacts.py
@@ -238,3 +238,16 @@ broadimpact_grouping_dict = {
         "truncating_plus": ["nonsense", "essential_splice", "splice_region_variant"],
         "nonsynonymous_splice": ["missense", "nonsense", "essential_splice"]
     }
+
+
+broadimpact_grouping_dict_with_synonymous = {
+        "missense": ["missense"],
+        "nonsense": ["nonsense"],
+        "essential_splice": ["essential_splice"],
+        "splice_region_variant": ["splice_region_variant"],
+        "truncating": ["nonsense", "essential_splice"],
+        "essential_splice_plus": ["essential_splice", "splice_region_variant"],
+        "truncating_plus": ["nonsense", "essential_splice", "splice_region_variant"],
+        "nonsynonymous_splice": ["missense", "nonsense", "essential_splice"],
+        "synonymous": ["synonymous"]
+    }
