+++ conf/tools/mutdensity.config
@@ -27,11 +27,11 @@ process {
     }
 
     withName: 'SUBSETMUTDENSITYADJUSTED' {
-        ext.filters     = ''
+        ext.filters     = { ['"TYPE": "SNV"'].join(',\t').trim()}
 
         ext.output_fmt  = { ['"header": true',
                                 '"columns": ["SAMPLE_ID",  "MUT_ID", "ALT_DEPTH", "canonical_SYMBOL", "canonical_Consequence_broader", "TYPE"]',
-                                '"colnames": ["SAMPLE_ID", "MUT_ID", "ALT_DEPTH", "SYMBOL", "canonical_Consequence_broader", "TYPE"]'
+                                '"colnames": ["SAMPLE_ID", "MUT_ID", "ALT_DEPTH", "GENE", "IMPACT", "TYPE"]'
                             ].join(',\t').trim()
                         }
