+++ modules/local/mut_density/main.nf
@@ -12,13 +12,13 @@ process MUTATION_DENSITY {
 
 
     output:
-    tuple val(meta), path("*.mutdensities.tsv") ,    emit: mutdensities
-    path("*.logfoldchangeplot.pdf") ,                emit: mutdensities_plots
+    tuple val(meta), path("*.mutdensities.tsv")         , emit: mutdensities
+    tuple val(meta), path("*.mutdensities_flat.tsv")    , emit: mutdensities_flat   
+    path("*.logfoldchangeplot.pdf")                     , emit: mutdensities_plots
     path "versions.yml" ,                            topic: versions
 
     script:
     def sample_name = "${meta.id}"
-    def panel_version = task.ext.panel_version ?: "${meta2.id}"
     """
     mut_density.py \\
                         --sample_name ${sample_name} \\
